Move TestSection class to separate file.
